bring over missing power saver blocker and other functionality from useChatEngine
