+++ ui/desktop/src/components/sessions/SessionHistoryView.tsx
@@ -33,7 +33,7 @@ import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/Tooltip';
 import { Message, Session } from '../../api';
 import { useNavigation } from '../../hooks/useNavigation';
 
-// Helper function to determine if a message is a user message (same as useChatEngine)
+// Helper function to determine if a message is a user message
 const isUserMessage = (message: Message): boolean => {
   if (message.role === 'assistant') {
     return false;
